+++ packages/aws-lambda/test/async/lambda.js
@@ -14,16 +14,37 @@ const fetch = require('node-fetch');
 
 const config = require('../../../serverless/test/config');
 
+const response = {
+  headers: {
+    'x-custom-header': 'custom header value'
+  },
+  body: {
+    message: 'Stan says hi!'
+  }
+};
+
+if (process.env.SERVER_TIMING_HEADER) {
+  if (process.env.SERVER_TIMING_HEADER === 'string') {
+    response.headers['sErveR-tIming'] = 'cache;desc="Cache Read";dur=23.2';
+  } else if (process.env.SERVER_TIMING_HEADER === 'array') {
+    response.multiValueHeaders = {
+      'ServEr-TiminG': ['cache;desc="Cache Read";dur=23.2', 'cpu;dur=2.4']
+    };
+  } else {
+    throw new Error(`Unknown SERVER_TIMING_HEADER value: ${process.env.SERVER_TIMING_HEADER}.`);
+  }
+}
+
 const handler = async event => {
   console.log('in actual handler');
   await fetch(config.downstreamDummyUrl);
   if (event.error) {
     throw new Error('Boom!');
   } else {
-    return {
-      statusCode: event.requestedStatusCode ? parseInt(event.requestedStatusCode, 10) : undefined,
-      message: 'Stan says hi!'
-    };
+    if (event.requestedStatusCode) {
+      response.statusCode = parseInt(event.requestedStatusCode, 10);
+    }
+    return response;
   }
 };