+++ packages/aws-lambda/test/callback/lambda.js
@@ -14,6 +14,27 @@ const http = require('http');
 
 const config = require('../../../serverless/test/config');
 
+const response = {
+  headers: {
+    'x-custom-header': 'custom header value'
+  },
+  body: {
+    message: 'Stan says hi!'
+  }
+};
+
+if (process.env.SERVER_TIMING_HEADER) {
+  if (process.env.SERVER_TIMING_HEADER === 'string') {
+    response.headers['sErveR-tIming'] = 'cache;desc="Cache Read";dur=23.2';
+  } else if (process.env.SERVER_TIMING_HEADER === 'array') {
+    response.multiValueHeaders = {
+      'ServEr-TiminG': ['cache;desc="Cache Read";dur=23.2', 'cpu;dur=2.4']
+    };
+  } else {
+    throw new Error(`Unknown SERVER_TIMING_HEADER value: ${process.env.SERVER_TIMING_HEADER}.`);
+  }
+}
+
 const handler = function handler(event, context, callback) {
   console.log('in actual handler');
   const req = http.get(config.downstreamDummyUrl, res => {
@@ -22,10 +43,10 @@ const handler = function handler(event, context, callback) {
       if (event.error) {
         callback(new Error('Boom!'));
       } else {
-        callback(null, {
-          statusCode: event.requestedStatusCode ? parseInt(event.requestedStatusCode, 10) : undefined,
-          message: 'Stan says hi!'
-        });
+        if (event.requestedStatusCode) {
+          response.statusCode = parseInt(event.requestedStatusCode, 10);
+        }
+        callback(null, response);
       }
     });
   });