+++ packages/serverless/src/index.js
@@ -4,3 +4,4 @@ exports.backendConnector = require('./backend_connector');
 exports.consoleLogger = require('./console_logger');
 exports.constants = require('./constants');
 exports.environment = require('./environment');
+exports.headers = require('./headers');